move sleep function to helpers
