+++ frontend/src/services/graphQL.ts
@@ -2,7 +2,7 @@ import axios, { AxiosResponse } from 'axios'
 import { getGraphQLApi, getTestHeader } from '../helpers/apiHelper'
 import { getToken } from './remoteit'
 import { store } from '../store'
-import sleep from './sleep'
+import sleep from '../helpers/sleep'
 
 let errorCount = 0
 const CLIENT_DEPRECATED = '121'
